Conflict resolutions (kept the parallel-exec architecture; adopted main's
post-merge-base changes where they don't touch it):
- exec/state.go: keep WorkerContext.RunTxTask; drop main's re-added Worker.Run
  loop / PublishReadMetrics (that worker model was replaced by the parallel
  scheduler on this branch).
- exec3_serial.go: drop the added PublishReadMetrics call on the deprecated
  serial path rather than add the method to the parallel worker.
- intra_block_state.go: keep our []types.Logs + logSize model; adapt
  ResumeLogIndexAt/ResetLogs (main's logArena additions) to it.
- domain_shared.go: adopt main's nonExecMetrics/commitmentNanos fields (used by
  merged code) and MergeMetrics(source,...); keep our DomainReader/AsGetter API.
- hex_patricia_hashed.go: keep error-checked updateCell; drop the removed
  mxTrieProcessedKeys hot-path counter.
- node/ethconfig/config.go: union (ExperimentalParallelCommitment + SlowBlockThreshold).
- writeset_normalize_bench_test.go: keep deleted (WriteSet.Normalize was removed).

+++ rpc/jsonrpc/trace_filtering.go
@@ -827,7 +827,7 @@ func (api *TraceAPIImpl) callBlock(
 			}
 		}
 		var amountWei uint256.Int
-		amountWei.Mul(uint256.NewInt(w.Amount), uint256.NewInt(common.GWei))
+		amountWei.Mul(uint256.NewInt(uint64(w.Amount)), uint256.NewInt(common.GWei))
 		wdiffs = append(wdiffs, withdrawalBalanceDiff{
 			address: w.Address,
 			prev:    prev,
